Update getCustomHeaders function
This change updates the getCustomHeaders helper function so that it returns an
associative array of the custom headers that we set via the system config. We
were never going to need the serialised string and we would have had to unpack
it every time we call the function anyway.

Also updated all invocations of the function accordingly.